Shader - add() - additional optional sources
Shader.add() allows to add shader sources of a specific type from a
file to a shader program.
The add() function now takes an additional optional parameter,
`optSrc`, which allows to append source code to the file's source
code after the `#version` line (first line).
For example, it can be used to append preprocessor directives.